Optimise `gamma::compress_u8` some more
Firstly, remove `f32::clamp` call and instead use `min` and `max` in
the branches where it’s needed.

Secondly, distribute the ‘+ 0.5’ to the branches.  This allows the
addition to be fused with the `mul_add` already present in the
non-linear branch.

This speeds up functions by around 5% (at least according to the
benchmarks):

    expand 8-bit    time:   [1.4407 us 1.4411 us 1.4416 us]
                    change: [-4.7316% -4.6949% -4.6611%] (p = 0.00 < 0.05)
    compress 8-bit  time:   [1.8470 us 1.8474 us 1.8478 us]
                    change: [-5.7560% -5.7234% -5.6839%] (p = 0.00 < 0.05)

(To be honest I’m not sure why expand_u8 sped up, but I’m gonna take it).