Add more gamma tests
First of all, add two more test cases in the `CASESS` array which cover
the linear part of the gamma function.  So far the only existing case
was for zero which is not very exciting.

Second of all, add a round-trip tests to make sure that the conversions
are inverses of each other.