Calculate gamma lookup table using rug crate
Use GNU MP (or more precisely the rug crate which provides bindings for
that library) for calculating the gamma lookup table.  This allows for
ridiculously and unnecessarily big precision which goes beyond even f64
or f128.  While this is an overkill, it may be useful if we ever start
providing f64 constants and functions.