Remove no-fma and prefer-fma features; release as 0.2.2
The two features were hastily defined and have two issues.

First of all, they are mutually exclusive which in general is not
a good thing for Rust crate features.  Second of all, they are defined
in terms of the fma instruction while that may not be what user really
cares about.  Rather, the features should be expressed in terms of
wanting fast, stable or accurate results.

Since the features are currently unused (as far as I can tell), delete
them for now and revisit when actual need to use them arises.