Support rgb crate by changing functions to accept `impl Into<[T; 3]>`
The rgb crate provides RGB struct for representing (red, green, blue)
tuples.  The struct can be converted from and into `[T; 3]` and since
srgb already operates on three-element arrays a lot of the times it’s
not too difficult to support RGB struct.

To that effect, change functions which took red, green and blue
components as a three-element array to be generic and accept any type
which can be converted into a three-element array.  Since RGB struct
can, it can be passed directly to those functions now.

The return value is unchanged and remain a three-element array.  Since
such arrays scan be converted into RGB struct this is probably good
enough.

The alternative would be to change return type (and perhaps also
argument type) to RGB but that would force users to depend on rgb
crate.  For the time being let’s not do that.