Define maths::sse module on all x86 targets
Currently the module is not defined unless target_feature="sse", however
sse::has_sse is called in matrix_product on all x86 targets even if SSE
target feature is not enabled.  This of course leads to failures when
building on for example i586.

This can be reproduced by building for i686 FreeBSD or OpenBSD; or using
i586-unknown-linux-gnu target on Linux.

Fix by defining maths::sse module on all x86 targets.  This means that
has_sse is present and matrix_product can use it.

Fixes: https://github.com/mina86/srgb.rs/issues/1